W2W_Py_Backend/requirements.txt
numpy == 1.16.2
opencv-python == 4.0.0.21
pafy == 0.5.4
youtube_dl == 2019.4.1
flask
flask-jsonpify
flask-sqlalchemy
flask-restful
tensorflow-gpu == 1.12.0
pillow == 5.4.1
scikit-image